pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pdf2audio"
version = "0.1.0"
description = "A simple, open-source tool that converts PDFs into audiobooks for free"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"
authors = [
    {name = "janakhpon", email = "janakhpon@example.com"}
]
keywords = ["pdf", "audio", "audiobook", "tts", "text-to-speech"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pypdf>=3.0.0",
    "pydub>=0.25.1",
    "kokoro-onnx>=0.3.4",
    "soundfile>=0.12.1",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
include = [
    "*.py",
    "LICENSE",
    "README.md",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
